requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
Appium-Python-Client==2.10.2
attrs==23.2.0
behave==1.2.6
certifi==2024.2.2
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
linecache2==1.0.0
outcome==1.3.0.post0
parse==1.20.1
parse-type==0.6.2
pluggy==1.4.0
PySocks==1.7.1
selenium==4.18.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
traceback2==1.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
unittest2==1.1.0
urllib3==2.2.1
wsproto==1.2.0
